[ALPS08631203] Add workaround for MediaTek MMIO timer
[Description]
The MT69XX series SoCs have the incomplete implementation issue in the
mmio timer. Specifically, the hardware only implements the TVAL
functionality, but not the CVAL functionality. This hardware limitation
will cause set_next_event_mem() fail to set the actual expiration time
when writing a value to the CVAL. On these platforms, the mmio timer's
internal expiration time will still be judged as 0 (the value of TVAL),
resulting in the mmio timer not functioning as intended.

The workaround is to use TVAL in addition to CVAL for these affected
platforms.

[Test]
Build pass and pass system reboot test

Signed-off-by: Walter Chang <walter.chang@mediatek.com>
CR-Id: ALPS08631203
Feature: RGU WDT Watchdog
Change-Id: I0fd449f001db23a2d14e20239dbacb3c9f2acc39